commit 20250514

+++ Core/Request/OficioNotificacionRequest.cs
@@ -2,7 +2,7 @@
 {
     public class OficioNotificacionRequest
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
         public string CiudadEnvio { get; set; }
         public DateTime FechaEnvio { get; set; }
         public string Membrete { get; set; }
@@ -10,9 +10,9 @@
         public string Ciudad { get; set; }
         public string Asunto { get; set; }
         public string Mensaje { get; set; }
-        public int IdAlertaSeguimiento { get; set; }
+        public long IdAlertaSeguimiento { get; set; }
         public string Comentario { get; set; }
-        public int IdNNA { get; set; }
+        public long IdNNA { get; set; }
         public string Cierre { get; set; }
         public string Firma { get; set; }
         public string FirmaJpg { get; set; }
